Fix C++ types

+++ Eratosthenes/_eratosthenes.cpp
@@ -130,7 +130,7 @@ std::vector<BigInteger> SieveOfEratosthenes(const BigInteger& n)
         return std::vector<BigInteger>(knownPrimes.begin(), highestPrimeIt);
     }
 
-    knownPrimes.reserve(std::expint(log(n)) - std::expint(log(2)));
+    knownPrimes.reserve(std::expint(log((double)n)) - std::expint(log(2)));
 
     // We are excluding multiples of the first few
     // small primes from outset. For multiples of
@@ -399,7 +399,7 @@ std::vector<BigInteger> SegmentedSieveOfEratosthenes(BigInteger n)
         return SieveOfEratosthenes(n);
     }
     std::vector<BigInteger> knownPrimes = SieveOfEratosthenes(limit);
-    knownPrimes.reserve(std::expint(log(n)) - std::expint(log(2)));
+    knownPrimes.reserve(std::expint(log((double)n)) - std::expint(log(2)));
 
     // Divide the range in different segments
     const size_t nCardinality = backward5(n);
@@ -509,7 +509,7 @@ BigInteger SegmentedCountPrimesTo(const BigInteger& n)
     const BigInteger practicalLimit = (sqrtnp1 < limit) ? sqrtnp1 : limit;
     std::vector<BigInteger> knownPrimes = SieveOfEratosthenes(practicalLimit);
     if (practicalLimit < sqrtnp1) {
-        knownPrimes.reserve(std::expint(log(sqrtnp1)) - std::expint(log(2)));
+        knownPrimes.reserve(std::expint(log((double)sqrtnp1)) - std::expint(log(2)));
     }
     size_t count = knownPrimes.size();
